+++ src/ElectroMagn/ElectroMagn2D.cpp
@@ -145,7 +145,7 @@ isNorthern(patch->isNorthern())
                         bufsize[i][isDual]--;
                     else if  (isDual) {
                         bufsize[i][isDual]--;
-                        if ( (patch->Pcoordinates[i]!=0) && (patch->Pcoordinates[i]!=params.number_of_patches[i]-1) ) 
+                        if ( (patch->Pcoordinates[i]!=0) && (patch->Pcoordinates[i]!=(unsigned int)params.number_of_patches[i]-1) ) 
                             bufsize[i][isDual]--;
                     }
